Update Localizer procedure after ICP
This commit changes the way we look for a better local map.

We introduce the concept of a neighbor local map : one that has all
vertices of the current local map but one, that is replaced by a
adjacent vertex of the current local map. If multiple adjacent
vertices exist, we select the one that is closest to the current robot
position.

Also, this commit removes the methods to find local maps from
MapManager, adding them to Localizer.